+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
 
 * Unreleased
+* 1.8 (2020-09-04)
+    * Auto-detect the location of 'auniter.ini' in the following order:
+      `--config` flag, the current directory, any parent directory,
+      `$HOME/auniter.ini`, and finally `$HOME/.auniter.ini`.
+    * Add `config` command to print the auto-detected `auniter.ini` file.
+    * Add `compile` command as an alias for `verify`, because the Arduino-CLI
+      binary uses `compile` instead of `verify`.
+    * Add `mon` command as an alias for `monitor`, because we often want
+      to run the 'monitor' command after an 'upmon', and this makes it
+      easier to retrieve the previous 'upmon' command from the bash history and
+      just remove the 'up'.
+    * Add support for [arduino-cli](https://github.com/arduino/arduino-cli)
+      using the `AUNITER_ARDUINO_CLI` environment variable. The `preprocessor`
+      directive in the `auniter.ini` must not contain a string due to a bug in
+      `arduino-cli`.
+    * Add better support and testing on MacOS 10.14 (Mojave).
 * 1.7.2 (2020-08-21)
     * Look for a `*.ino` file in the current directory if no sketch file is
       specified for auniter.sh.